refactor: rename SlotTemplator to ComponentTemplator and update related functionality
