+++ tests/View/Templator/view/componentcard.template
@@ -1,0 +1,1 @@
+<div>{% yield('componentcard.template') %}</div>
