--- tests/View/Templator/view/slot.template
@@ -1,1 +1,0 @@
-<html><head></head><body>{% yield('slot.template') %}</body></html>
