--- tests/View/Templator/view/slotyield.template
@@ -1,1 +1,0 @@
-<html><head></head><body>{% yield('slot2.template') %}</body></html>
